feat: cell_evaluation is moved to library
