+++ ai3_negaalpha.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include "src/lib/board.hpp"
-#include "cell_evaluation.hpp"
+#include "src/lib/cell_evaluation.hpp"
 
 using namespace std;
