+++ src/lib/cell_evaluation.cpp
